Rename module `decon` -> `sexp`.
