+++ src/collections/package.json
@@ -15,7 +15,7 @@
   },
 
   "dependencies": {
-    "@this/decon": "*",
+    "@this/sexp": "*",
     "@this/typey": "*"
   }
 }
