+++ src/net-util/export/DispatchInfo.js
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { PathKey } from '@this/collections';
-import { IntfDeconstructable, Sexp } from '@this/decon';
+import { IntfDeconstructable, Sexp } from '@this/sexp';
 import { IntfLogger } from '@this/loggy-intf';
 import { MustBe } from '@this/typey';
