+++ src/net-util/package.json
@@ -16,7 +16,7 @@
   "dependencies": {
     "@this/async": "*",
     "@this/collections": "*",
-    "@this/decon": "*",
+    "@this/sexp": "*",
     "@this/fs-util": "*",
     "@this/loggy-intf": "*",
     "@this/quant": "*",
