+++ src/sexp/export/IntfDeconstructable.js
