+++ src/sexp/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@this/decon",
+  "name": "@this/sexp",
   "version": "0.0.1",
   "type": "module",
   "private": true,
