+++ src/valvis/export/ErrorUtil.js
@@ -1,7 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { IntfDeconstructable } from '@this/decon';
+import { IntfDeconstructable } from '@this/sexp';
 
 import { StackTrace } from '#x/StackTrace';
